refactor: split gRPC API frontend
