+++ Cargo.toml
@@ -26,6 +26,7 @@ default = []
 enterprise = [
     "audit/enterprise",
     "openobserve-api/enterprise",
+    "openobserve-api-grpc/enterprise",
     "openobserve-core/enterprise",
     "openobserve-jobs/enterprise",
     "super_cluster_queue/enterprise",
@@ -35,6 +36,7 @@ enterprise = [
 ]
 vectorscan = [
     "openobserve-api/vectorscan",
+    "openobserve-api-grpc/vectorscan",
     "openobserve-core/vectorscan",
     "openobserve-jobs/vectorscan",
     "super_cluster_queue/vectorscan",
@@ -42,6 +44,7 @@ vectorscan = [
 ]
 cloud = [
     "openobserve-api/cloud",
+    "openobserve-api-grpc/cloud",
     "openobserve-core/cloud",
     "openobserve-jobs/cloud",
 ]
@@ -219,6 +222,7 @@ async-stream.workspace = true
 sourcemap.workspace = true
 audit.workspace = true
 openobserve-api.workspace = true
+openobserve-api-grpc.workspace = true
 web.workspace = true
 openobserve-api-management.workspace = true
 openobserve-core.workspace = true
@@ -253,6 +257,7 @@ transform.workspace = true
 [workspace]
 members = [
     "src/api",
+    "src/api_grpc",
     "src/api_management",
     "src/api_query",
     "src/api_common",
@@ -330,6 +335,7 @@ audit = { path = "src/audit" }
 promql = { path = "src/promql" }
 promql-service = { path = "src/promql_service" }
 openobserve-api = { path = "src/api" }
+openobserve-api-grpc = { path = "src/api_grpc" }
 openobserve-api-management = { path = "src/api_management" }
 openobserve-api-query = { path = "src/api_query" }
 openobserve-api-common = { path = "src/api_common" }
@@ -433,7 +439,7 @@ datafusion-proto = "54"
 datafusion-functions-aggregate-common = "54"
 datafusion-functions-json = { git = "https://github.com/openobserve/datafusion-functions-json", rev = "0df53d710425cc91cf109e77a050bbe1e374e4fe" }
 expect-test = "1.4"
-arrow = { version = "58", features = ["ipc_compression", "prettyprint"] }
+arrow = { version = "58", features = ["ipc_compression"] }
 arrow-flight = "58"
 arrow-json = "58"
 arrow-schema = { version = "58", features = ["serde"] }
@@ -580,10 +586,8 @@ utoipa-swagger-ui = { version = "9", features = ["axum", "vendored"] }
 uuid = { version = "1.20", features = ["v7"] }
 vector-enrichment = { version = "0.1.0", package = "enrichment", git = "https://github.com/openobserve/vector", branch = "branch-v0.54" }
 vrl = { version = "0.31", features = [
-    "arbitrary",
     "value",
     "compiler",
-    "test",
 ] }
 x509-parser = "0.18.0"
 zip = "8.0"
