+++ src/api/src/lib.rs
@@ -23,3 +23,4 @@ use openobserve_core as service;
 
 pub mod handler;
 pub mod router;
+pub mod server;
