+++ src/api/src/router/mod.rs
@@ -13,7 +13,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod grpc;
 pub mod http;
 #[cfg(feature = "enterprise")]
 pub mod ratelimit;
