+++ src/api_grpc/src/handler/mod.rs
@@ -13,4 +13,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod http;
+pub mod grpc;
