+++ src/api_grpc/src/lib.rs
@@ -13,4 +13,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod http;
+//! gRPC transport frontend for OpenObserve.
+
+use common;
+use openobserve_core as service;
+
+pub mod handler;
+pub mod router;
+pub mod server;
