+++ src/core/Cargo.toml
@@ -71,7 +71,6 @@ cron.workspace = true
 dashmap.workspace = true
 datafusion.workspace = true
 arrow.workspace = true
-arrow-json.workspace = true
 arrow-schema.workspace = true
 flate2.workspace = true
 futures.workspace = true
@@ -85,8 +84,6 @@ o2_enterprise = { workspace = true, optional = true, default-features = false }
 o2_openfga = { workspace = true, optional = true }
 opentelemetry.workspace = true
 opentelemetry-proto.workspace = true
-parking_lot.workspace = true
-prometheus.workspace = true
 promql-parser.workspace = true
 prost.workspace = true
 prost-wkt-types.workspace = true
@@ -114,7 +111,3 @@ infra.workspace = true
 ingester.workspace = true
 chromiumoxide.workspace = true
 lettre.workspace = true
-futures-util.workspace = true
-
-[dev-dependencies]
-multer.workspace = true
